+++ tests/trial_tests/test_trial.py
@@ -16,6 +16,7 @@ from optuna import distributions
 from optuna import load_study
 from optuna import samplers
 from optuna import storages
+from optuna import Study
 from optuna.distributions import CategoricalDistribution
 from optuna.distributions import FloatDistribution
 from optuna.distributions import IntDistribution
@@ -24,7 +25,9 @@ from optuna.testing.samplers import DeterministicSampler
 from optuna.testing.storages import STORAGE_MODES
 from optuna.testing.storages import StorageSupplier
 from optuna.testing.tempfile_pool import NamedTemporaryFilePool
+from optuna.trial import FrozenTrial
 from optuna.trial import Trial
+from optuna.trial import TrialState
 from optuna.trial._trial import _LazyTrialSystemAttrs
 
 
@@ -745,3 +748,24 @@ def test_suggest_int_positional_args(positional_args_names: list[str]) -> None:
     args = [kwargs[name] for name in positional_args_names]
     # No error should not be raised even if the coding style is old.
     trial.suggest_int("x", -1, 1, *args)
+
+
+def test_before_trial() -> None:
+    n_calls = 0
+    n_trials = 3
+
+    class SamplerBeforeTrial(optuna.samplers.RandomSampler):
+        def before_trial(self, study: Study, trial: FrozenTrial) -> None:
+            assert len(study.trials) - 1 == trial.number
+            assert trial.state == TrialState.RUNNING
+            assert trial.values is None
+            nonlocal n_calls
+            n_calls += 1
+
+    sampler = SamplerBeforeTrial()
+    study = optuna.create_study(directions=["minimize", "minimize"], sampler=sampler)
+
+    study.optimize(
+        lambda t: [t.suggest_float("y", -3, 3), t.suggest_int("x", 0, 10)], n_trials=n_trials
+    )
+    assert n_calls == n_trials
